Velocities are all i16s

+++ sc_common/src/net/cb.rs
@@ -67,7 +67,7 @@ pub enum Packet {
   },
   /// Change of an entity's velocity.
   #[id = 9]
-  EntityVelocity { eid: i32, x: i32, y: i32, z: i32 },
+  EntityVelocity { eid: i32, x: i16, y: i16, z: i16 },
   #[id = 10]
   JoinGame {
     eid:                   i32,
@@ -133,9 +133,9 @@ pub enum Packet {
     yaw:      i8,
     pitch:    i8,
     head_yaw: i8,
-    vel_x:    i32,
-    vel_y:    i32,
-    vel_z:    i32,
+    vel_x:    i16,
+    vel_y:    i16,
+    vel_z:    i16,
   },
   #[id = 17]
   SpawnPlayer { eid: i32, id: UUID, x: f64, y: f64, z: f64, yaw: i8, pitch: i8 },
